Refactor invoice creation process
