--- spec/jobs/send_mail_job_spec.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-RSpec.describe SendMailJob, type: :job do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
