--- spec/jobs/send_siat_job_spec.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-RSpec.describe SendSiatJob, type: :job do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
